+++ news.md
@@ -9,7 +9,13 @@
 * `bqSaveSchema()`, `bqExtractSchema()` - allow to save schema from a given data set into a JSON file
 
 * `bqDeleteDataset()` - before deletion presense of `delete:never` label key-value pair is checked, 
-    which will protect datasets from programmatic deletion.
+    which will protect datasets from programmatic deletion. (#79)
+    
+* `bqProjectDatasets()` - lists datasets in the project. (#79)
+    
+* `bqProjectTables()` - extracts metadata for all tables in the project by extracting `__TABLES__` for each dataset. (#79)
+
+* `bqUseLegacySql()` - allows to check or set flavour of bigquery sql. (#79)
 
 # retl 0.1.1